+++ accounting/src/funcs/accountsCreate.ts
@@ -3,16 +3,14 @@
  */
 
 import { AccountingCore } from "../core.js";
-import {
-  encodeFormQuery as encodeFormQuery$,
-  encodeJSON as encodeJSON$,
-  encodeSimple as encodeSimple$,
-} from "../lib/encodings.js";
-import * as m$ from "../lib/matchers.js";
-import * as schemas$ from "../lib/schemas.js";
+import { encodeFormQuery, encodeJSON, encodeSimple } from "../lib/encodings.js";
+import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
+import { AccountingError } from "../sdk/models/errors/accountingerror.js";
 import {
   ConnectionError,
   InvalidRequestError,
@@ -20,10 +18,11 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../sdk/models/errors/httpclienterrors.js";
-import { SDKError } from "../sdk/models/errors/sdkerror.js";
+import { ResponseValidationError } from "../sdk/models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../sdk/models/errors/sdkvalidationerror.js";
 import * as operations from "../sdk/models/operations/index.js";
 import * as shared from "../sdk/models/shared/index.js";
+import { APICall, APIPromise } from "../sdk/types/async.js";
 import { Result } from "../sdk/types/fp.js";
 
 /**
@@ -40,97 +39,116 @@ import { Result } from "../sdk/types/fp.js";
  *
  * Check out our [coverage explorer](https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=chartOfAccounts) for integrations that support creating an account.
  */
-export async function accountsCreate(
-  client$: AccountingCore,
+export function accountsCreate(
+  client: AccountingCore,
   companyId: string,
   connectionId: string,
-  accountPrototype?: shared.AccountPrototype | undefined,
   timeoutInMinutes?: number | undefined,
+  accountPrototype?: shared.AccountPrototype | undefined,
   options?: RequestOptions,
-): Promise<
+): APIPromise<
   Result<
     operations.CreateAccountResponse,
-    | SDKError
-    | SDKValidationError
-    | UnexpectedClientError
-    | InvalidRequestError
+    | AccountingError
+    | ResponseValidationError
+    | ConnectionError
     | RequestAbortedError
     | RequestTimeoutError
-    | ConnectionError
+    | InvalidRequestError
+    | UnexpectedClientError
+    | SDKValidationError
   >
 > {
-  const input$: operations.CreateAccountRequest = {
-    accountPrototype: accountPrototype,
+  return new APIPromise($do(
+    client,
+    companyId,
+    connectionId,
+    timeoutInMinutes,
+    accountPrototype,
+    options,
+  ));
+}
+
+async function $do(
+  client: AccountingCore,
+  companyId: string,
+  connectionId: string,
+  timeoutInMinutes?: number | undefined,
+  accountPrototype?: shared.AccountPrototype | undefined,
+  options?: RequestOptions,
+): Promise<
+  [
+    Result<
+      operations.CreateAccountResponse,
+      | AccountingError
+      | ResponseValidationError
+      | ConnectionError
+      | RequestAbortedError
+      | RequestTimeoutError
+      | InvalidRequestError
+      | UnexpectedClientError
+      | SDKValidationError
+    >,
+    APICall,
+  ]
+> {
+  const input: operations.CreateAccountRequest = {
     companyId: companyId,
     connectionId: connectionId,
     timeoutInMinutes: timeoutInMinutes,
+    accountPrototype: accountPrototype,
   };
 
-  const parsed$ = schemas$.safeParse(
-    input$,
-    (value$) => operations.CreateAccountRequest$outboundSchema.parse(value$),
+  const parsed = safeParse(
+    input,
+    (value) => operations.CreateAccountRequest$outboundSchema.parse(value),
     "Input validation failed",
   );
-  if (!parsed$.ok) {
-    return parsed$;
+  if (!parsed.ok) {
+    return [parsed, { status: "invalid" }];
   }
-  const payload$ = parsed$.value;
-  const body$ = encodeJSON$("body", payload$.accountPrototype, {
-    explode: true,
-  });
+  const payload = parsed.value;
+  const body = encodeJSON("body", payload.accountPrototype, { explode: true });
 
-  const pathParams$ = {
-    companyId: encodeSimple$("companyId", payload$.companyId, {
+  const pathParams = {
+    companyId: encodeSimple("companyId", payload.companyId, {
       explode: false,
       charEncoding: "percent",
     }),
-    connectionId: encodeSimple$("connectionId", payload$.connectionId, {
+    connectionId: encodeSimple("connectionId", payload.connectionId, {
       explode: false,
       charEncoding: "percent",
     }),
   };
 
-  const path$ = pathToFunc(
+  const path = pathToFunc(
     "/companies/{companyId}/connections/{connectionId}/push/accounts",
-  )(pathParams$);
+  )(pathParams);
 
-  const query$ = encodeFormQuery$({
-    "timeoutInMinutes": payload$.timeoutInMinutes,
+  const query = encodeFormQuery({
+    "timeoutInMinutes": payload.timeoutInMinutes,
   });
 
-  const headers$ = new Headers({
+  const headers = new Headers(compactMap({
     "Content-Type": "application/json",
     Accept: "application/json",
-  });
+  }));
 
-  const authHeader$ = await extractSecurity(client$.options$.authHeader);
-  const security$ = authHeader$ == null ? {} : { authHeader: authHeader$ };
+  const secConfig = await extractSecurity(client._options.authHeader);
+  const securityInput = secConfig == null ? {} : { authHeader: secConfig };
+  const requestSecurity = resolveGlobalSecurity(securityInput);
+
   const context = {
+    options: client._options,
+    baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "create-account",
     oAuth2Scopes: [],
-    securitySource: client$.options$.authHeader,
-  };
-  const securitySettings$ = resolveGlobalSecurity(security$);
 
-  const requestRes = client$.createRequest$(context, {
-    security: securitySettings$,
-    method: "POST",
-    path: path$,
-    headers: headers$,
-    query: query$,
-    body: body$,
-    timeoutMs: options?.timeoutMs || client$.options$.timeoutMs || -1,
-  }, options);
-  if (!requestRes.ok) {
-    return requestRes;
-  }
-  const request$ = requestRes.value;
+    resolvedSecurity: requestSecurity,
 
-  const doResult = await client$.do$(request$, {
-    context,
-    errorCodes: [],
+    securitySource: client._options.authHeader,
     retryConfig: options?.retries
-      || client$.options$.retryConfig
+      || client._options.retryConfig
       || {
         strategy: "backoff",
         backoff: {
@@ -140,15 +158,39 @@ export async function accountsCreate(
           maxElapsedTime: 3600000,
         },
         retryConnectionErrors: true,
-      },
+      }
+      || { strategy: "none" },
     retryCodes: options?.retryCodes || ["408", "429", "5XX"],
+  };
+
+  const requestRes = client._createRequest(context, {
+    security: requestSecurity,
+    method: "POST",
+    baseURL: options?.serverURL,
+    path: path,
+    headers: headers,
+    query: query,
+    body: body,
+    userAgent: client._options.userAgent,
+    timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
+  }, options);
+  if (!requestRes.ok) {
+    return [requestRes, { status: "invalid" }];
+  }
+  const req = requestRes.value;
+
+  const doResult = await client._do(req, {
+    context,
+    errorCodes: [],
+    retryConfig: context.retryConfig,
+    retryCodes: context.retryCodes,
   });
   if (!doResult.ok) {
-    return doResult;
+    return [doResult, { status: "request-error", request: req }];
   }
   const response = doResult.value;
 
-  const responseFields$ = {
+  const responseFields = {
     ContentType: response.headers.get("content-type")
       ?? "application/octet-stream",
     StatusCode: response.status,
@@ -156,28 +198,32 @@ export async function accountsCreate(
     Headers: {},
   };
 
-  const [result$] = await m$.match<
+  const [result] = await M.match<
     operations.CreateAccountResponse,
-    | SDKError
-    | SDKValidationError
-    | UnexpectedClientError
-    | InvalidRequestError
+    | AccountingError
+    | ResponseValidationError
+    | ConnectionError
     | RequestAbortedError
     | RequestTimeoutError
-    | ConnectionError
+    | InvalidRequestError
+    | UnexpectedClientError
+    | SDKValidationError
   >(
-    m$.json(200, operations.CreateAccountResponse$inboundSchema, {
+    M.json(200, operations.CreateAccountResponse$inboundSchema, {
       key: "CreateAccountResponse",
     }),
-    m$.json(
-      [400, 401, 402, 403, 404, 429, 500, 503],
+    M.json(
+      [400, 401, 402, 403, 404, 429],
       operations.CreateAccountResponse$inboundSchema,
       { key: "ErrorMessage" },
     ),
-  )(response, { extraFields: responseFields$ });
-  if (!result$.ok) {
-    return result$;
+    M.json([500, 503], operations.CreateAccountResponse$inboundSchema, {
+      key: "ErrorMessage",
+    }),
+  )(response, req, { extraFields: responseFields });
+  if (!result.ok) {
+    return [result, { status: "complete", request: req, response }];
   }
 
-  return result$;
+  return [result, { status: "complete", request: req, response }];
 }