+++ lending/src/models/errors/index.ts
@@ -4,5 +4,6 @@
 
 export * from "./errormessage.js";
 export * from "./httpclienterrors.js";
+export * from "./responsevalidationerror.js";
 export * from "./sdkerror.js";
 export * from "./sdkvalidationerror.js";